StickmanFactory/requirements.txt
svgwrite>=1.4.3
jsonschema>=4.17.0
soundfile>=0.12.1
numpy>=1.24.0
tqdm>=4.65.0
Pillow>=10.0.0
